refactor(Runtime): add base class for dynamic views which `AppView` extends
`DynamicView` will eventually serve as the base type only for views that
supported being created and destroyed dynamically (i.e. embedded and host
views). However, for a transitional period, `AppView` (and thus all generated
views) will extend this class, as `AppView` already implements its semantics
today.

PiperOrigin-RevId: 242534841
